Add experimental multigraph debugger
The existing cog visual debugger is limited to handling a single
compute graph/model.  As we look to add distributed computation to
the CCT, we need tools that can support multiple graphs. This commit
introduces a simple prototype debugger, based on the existing
version, that has support for muliple graphs and parameter servers.

It differs primarily in that the probe desktop area is tabbed, one
tab per graph, so that probes attached to different graphs can be
viewed in a single debugger instance. It is limited in that probe
views from multiple graphs cannot be viewed side by side due to the
tab interface, and it assumes that all graphs being debugged are
identical in structure and differ only in state.