Add simple parameter server prototype
This commit adds a barebones parameter server implementation and a
simple  test harness for debugging it with the experimental
multigraph debugger.

The parameter server is intended only to collect and distribute a
backpropagation signal; it does not implement any learning rule. A
wrapper for existing learning rules is included to act as an adapter
to the parameter server.