Support unpipelined sensors/actuators in param server rule
Added extra flags to ParameterServerLearningRule's constructor to
enable/disable pipelining for the sensor and actuator that
communicate with the parameter server. Pipelining allows for more
throughput to the server, but also adds latency (additional pipeline
stages) that may not be tolerable for a given model.

Pipelining defaults to off for both sensors and actuators, which is
a change from how the learning rule worked before.