Add a toggleable check for NaNs to param serve push op
Some of the test cases being using to prototype the parameter server
are generating NaNs after a number of steps. This commit adds a
NaN check to the reference ParameterServerImpl's push(...) definition
(enabled by default) that will throw an error the moment it first
detects a NaN, to aid in debugging.