Renamed a testcase, generalized test harness
Renamed the TestParameterServerGraph to DistributedLogisticRegression,
to make it more apparent what the graph is actually doing and because
going forward we'll be exercising with the server implementation with
additional graphs.

Tweaked the test harness to make it easier to plug in different
graphs. It's no longer tied to the logistic regression example, i.e.
it's reusable. All the harness is responsible for now is launching the
debugging window.